[Wallet] Display price+total in wallet
Fixes:
-update price
